refactor hotkeys to hooks
