+++ packages/renderer/src/components/modules/IconsHome/IconCardsSection/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useHotKeyConfig';
