+++ packages/renderer/src/components/modules/IconsHome/IconCardsSection/index.tsx
@@ -1,12 +1,13 @@
 import type { Dispatch, FC, SetStateAction } from 'react';
+import { useEffect, useState } from 'react';
 import { useRef } from 'react';
 import type { Icon } from '/@/data/icons/types';
 import { HotKeys } from 'react-hotkeys';
 import { IconCard } from './IconCard';
 import { EmptyPlaceholder } from './EmptyPlaceholder';
 import { IconContextMenu } from './IconContextMenu';
-import { findSelectedIconPos, getNumberOfIconInRow } from './util';
 import { VirtualizedGrid } from '@mierak/react-virtualized-grid';
+import { useHotKeyConfig } from './hooks';
 
 interface Props {
   icons?: Icon[];
@@ -23,66 +24,17 @@ export const IconCardsSection: FC<Props> = ({
 }) => {
   const wrapperDivRef = useRef<HTMLDivElement>(null);
 
-  const selectIcon = (icon: Icon) => {
-    setSelectedIcon(icon);
-    const iconDom = document.querySelector(`[data-icon-card-id="${icon.id}"]`);
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    iconDom?.scrollIntoViewIfNeeded(false);
-  };
+  const { keyMap, handlers } = useHotKeyConfig({ icons, setSelectedIcon });
 
-  const keyMap = {
-    MOVE_UP: 'up',
-    MOVE_DOWN: 'down',
-    MOVE_RIGHT: 'right',
-    MOVE_LEFT: 'left',
-  };
+  const [refreshGrid, setRefreshGrid] = useState(false);
+  useEffect(() => {
+    setRefreshGrid(true);
+    setTimeout(() => setRefreshGrid(false), 1);
+  }, [icons]);
 
-  const handlers = {
-    MOVE_RIGHT: (keyEvent?: Event) => {
-      if (icons?.length) {
-        const selectedIconPos = findSelectedIconPos(icons);
-
-        if (typeof selectedIconPos === 'number' && icons[selectedIconPos + 1]) {
-          selectIcon(icons[selectedIconPos + 1]);
-        }
-
-        keyEvent?.preventDefault();
-      }
-    },
-    MOVE_LEFT: (keyEvent?: Event) => {
-      if (icons?.length) {
-        const selectedIconPos = findSelectedIconPos(icons);
-
-        if (typeof selectedIconPos === 'number' && icons[selectedIconPos - 1]) {
-          selectIcon(icons[selectedIconPos - 1]);
-        }
-        keyEvent?.preventDefault();
-      }
-    },
-    MOVE_DOWN: (keyEvent?: Event) => {
-      if (icons?.length) {
-        const selectedIconPos = findSelectedIconPos(icons);
-        const numberOfIconInRow = getNumberOfIconInRow();
-
-        if (typeof selectedIconPos === 'number' && icons[selectedIconPos + numberOfIconInRow]) {
-          selectIcon(icons[selectedIconPos + numberOfIconInRow]);
-        }
-        keyEvent?.preventDefault();
-      }
-    },
-    MOVE_UP: (keyEvent?: Event) => {
-      if (icons?.length) {
-        const selectedIconPos = findSelectedIconPos(icons);
-        const numberOfIconInRow = getNumberOfIconInRow();
-
-        if (typeof selectedIconPos === 'number' && icons[selectedIconPos - numberOfIconInRow]) {
-          selectIcon(icons[selectedIconPos - numberOfIconInRow]);
-        }
-        keyEvent?.preventDefault();
-      }
-    },
-  };
+  if (refreshGrid) {
+    return <></>;
+  }
 
   if (!icons?.length) {
     return <EmptyPlaceholder searchQuery={searchQuery} />;
@@ -105,14 +57,12 @@ export const IconCardsSection: FC<Props> = ({
             {(index) => {
               const icon = icons[index];
               return (
-                icon && (
-                  <IconCard
-                    key={icon?.id}
-                    icon={icon}
-                    isSelected={selectedIcon?.id === icon?.id}
-                    setSelectedIcon={setSelectedIcon}
-                  />
-                )
+                <IconCard
+                  key={icon?.id}
+                  icon={icon}
+                  isSelected={selectedIcon?.id === icon?.id}
+                  setSelectedIcon={setSelectedIcon}
+                />
               );
             }}
           </VirtualizedGrid>
